fix: add sol and jnlp console to converted host
