+++ containers/Compute/views/physicalmachine/components/List.vue
@@ -25,6 +25,7 @@ import ListMixin from '@/mixins/list'
 import { getDomainChangeOwnerAction, getSetPublicAction, getEnabledSwitchActions } from '@/utils/common/tableActions'
 import { hasServices } from '@/utils/auth'
 import regexp from '@/utils/regexp'
+import { hostCommonActions } from '../../../utils/hostActions'
 
 export default {
   name: 'PhysicalmachineList',
@@ -154,192 +155,7 @@ export default {
           label: this.$t('compute.text_275'),
           actions: (obj) => {
             return [
-              {
-                label: this.$t('compute.text_272'),
-                action: (obj) => {
-                  this.list.batchPerformAction('start', null, this.list.steadyStatus)
-                },
-                meta: () => {
-                  if (this.list.selectedItems.length <= 0) {
-                    return {
-                      validate: false,
-                      tooltip: this.$t('compute.text_823'),
-                    }
-                  }
-                  for (let i = 0; i < this.list.selectedItems.length; i++) {
-                    const obj = this.list.selectedItems[i]
-                    if (!obj.is_baremetal) {
-                      return {
-                        validate: false,
-                        tooltip: this.$t('compute.text_823'),
-                      }
-                    }
-                    if (obj.server_id && obj.host_type === 'baremetal') {
-                      return {
-                        validate: false,
-                      }
-                    }
-                    if (obj.status !== 'ready') {
-                      return {
-                        validate: false,
-                        tooltip: this.$t('compute.text_823'),
-                      }
-                    }
-                  }
-                  return {
-                    validate: true,
-                  }
-                },
-              },
-              {
-                label: this.$t('compute.text_273'),
-                action: (obj) => {
-                  this.list.batchPerformAction('stop', null, this.list.steadyStatus)
-                },
-                meta: () => {
-                  if (this.list.selectedItems.length <= 0) {
-                    return {
-                      validate: false,
-                      tooltip: this.$t('compute.text_824'),
-                    }
-                  }
-                  for (let i = 0; i < this.list.selectedItems.length; i++) {
-                    const obj = this.list.selectedItems[i]
-                    if (!obj.is_baremetal) {
-                      return {
-                        validate: false,
-                        tooltip: this.$t('compute.text_824'),
-                      }
-                    }
-                    if (obj.server_id && obj.host_type === 'baremetal') {
-                      return {
-                        validate: false,
-                      }
-                    }
-                    if (obj.status !== 'running') {
-                      return {
-                        validate: false,
-                        tooltip: this.$t('compute.text_824'),
-                      }
-                    }
-                  }
-                  return {
-                    validate: true,
-                  }
-                },
-              },
-              {
-                label: this.$t('compute.text_550'),
-                action: () => {
-                  this.list.batchPerformAction('maintenance', null, this.list.steadyStatus)
-                },
-                meta: () => {
-                  if (this.list.selectedItems.length <= 0) {
-                    return {
-                      validate: false,
-                    }
-                  }
-                  for (let i = 0; i < this.list.selectedItems.length; i++) {
-                    const obj = this.list.selectedItems[i]
-                    if (obj.server) {
-                      return {
-                        validate: false,
-                      }
-                    }
-                    if (!obj.is_baremetal) {
-                      return {
-                        validate: false,
-                      }
-                    }
-                    if (!obj.server_id) {
-                      return {
-                        validate: false,
-                      }
-                    }
-                    if (obj.is_maintenance) {
-                      return {
-                        validate: false,
-                      }
-                    }
-                    if (['running', 'ready'].indexOf(obj.status) < 0) {
-                      return {
-                        validate: false,
-                      }
-                    }
-                  }
-                  return {
-                    validate: true,
-                  }
-                },
-              },
-              {
-                label: this.$t('compute.text_559'),
-                action: () => {
-                  this.list.batchPerformAction('unmaintenance', null, this.list.steadyStatus)
-                },
-                meta: () => {
-                  if (this.list.selectedItems.length <= 0) {
-                    return {
-                      validate: false,
-                    }
-                  }
-                  for (let i = 0; i < this.list.selectedItems.length; i++) {
-                    const obj = this.list.selectedItems[i]
-                    if (obj.server) {
-                      return {
-                        validate: false,
-                      }
-                    }
-                    if (!obj.is_baremetal) {
-                      return {
-                        validate: false,
-                      }
-                    }
-                    if (!obj.server_id) {
-                      return {
-                        validate: false,
-                      }
-                    }
-                    if (!obj.is_maintenance) {
-                      return {
-                        validate: false,
-                      }
-                    }
-                    if (['running', 'ready'].indexOf(obj.status) < 0) {
-                      return {
-                        validate: false,
-                      }
-                    }
-                  }
-                  return {
-                    validate: true,
-                  }
-                },
-              },
-              {
-                label: this.$t('compute.perform_sync_status'),
-                action: () => {
-                  this.list.batchPerformAction('syncstatus', null, this.list.steadyStatus)
-                },
-                meta: () => {
-                  if (this.list.selectedItems.length <= 0) {
-                    return {
-                      validate: false,
-                    }
-                  }
-                  for (let i = 0; i < this.list.selectedItems.length; i++) {
-                    const obj = this.list.selectedItems[i]
-                    if (!obj.is_baremetal) {
-                      return {
-                        validate: false,
-                      }
-                    }
-                  }
-                  return {
-                    validate: true,
-                  }
-                },
-              },
+              ...hostCommonActions(this),
               getDomainChangeOwnerAction(this, {
                 name: this.$t('dictionary.host'),
                 resource: 'hosts',
